requirements/requirements-base.txt
bokeh==1.4.0
bottleneck==1.3.2
cvxpy==1.1.11
dask==2021.2.0
datashader==0.11.1
distributed==2021.2.0
ecos==2.0.7.post1
ffmpeg-python==0.2.0
holoviews==1.12.7
ipython==7.12.0
ipywidgets==7.5.1
jupyter==1.0.0
matplotlib==3.1.3
natsort==7.0.1
netcdf4==1.5.3
networkx==2.4
numba==0.52.0
numpy==1.18.1
opencv-python==4.2.0.34
pandas==1.0.1
panel==0.8.0
papermill==2.0.0
param==1.9.3
pip==20.0.2
pyfftw==0.12.0
pytest-cov==2.9.0
scikit-image==0.18.1
scikit-learn==0.22.1
scipy==1.4.1
scs==2.1.1-2
simpleitk==2.0.2
sk-video==1.1.10
statsmodels==0.11.1
tifffile==2020.2.16
tqdm==4.43.0
xarray==0.17.0
zarr==2.4.0
sparse==0.11.2
pymetis==2020.1
rechunker==0.3.3
medpy==0.4.0
